+++ toqito/state_metrics/bures_distance.py
@@ -8,16 +8,16 @@ def bures_distance(rho_1: np.ndarray, rho_2: np.ndarray, decimals: int = 10) -> 
     r"""
     Compute the Bures distance of two density matrices [WikBures]_.
 
-    Calculate the Bures distance between the two density matrices: :code:`rho_1` and
-    :code:`rho_2`, defined by:
+    Calculate the Bures distance between two density matrices :code:`rho_1` and :code:`rho_2`
+    defined by:
+
     .. math::
         \sqrt{2 (1 - F(\rho_1, \rho_2))},
 
-    where :math:`F(\cdot)` denotes the fidelity between :math:`\rho_1` and :math:`\rho_2`.
-    The return is a value between :math:`0` and :math:`\sqrt{2}`,with
-    :math:`0` corresponding to matrices: :code:`rho_1 = rho_2` and
-    :math:`\sqrt{2}` corresponding to the case: :code:`rho_1`and
-    :code:`rho_2` with orthogonal support.
+    where :math:`F(\cdot)` denotes the fidelity between :math:`\rho_1` and :math:`\rho_2`. The
+    return is a value between :math:`0` and :math:`\sqrt{2}`,with :math:`0` corresponding to
+    matrices: :code:`rho_1 = rho_2` and :math:`\sqrt{2}` corresponding to the case: :code:`rho_1`
+    and :code:`rho_2` with orthogonal support.
 
     Examples
     ==========
@@ -65,5 +65,5 @@ def bures_distance(rho_1: np.ndarray, rho_2: np.ndarray, decimals: int = 10) -> 
     # Perform some error checking.
     if not np.all(rho_1.shape == rho_2.shape):
         raise ValueError("InvalidDim: `rho_1` and `rho_2` must be matrices of the same size.")
-    # Round fidelity to only 10 decimals to avoid error when rho_1 = rho_2
+    # Round fidelity to only 10 decimals to avoid error when :code:`rho_1 = rho_2`.
     return np.sqrt(2.0 * (1.0 - np.round(fidelity(rho_1, rho_2), decimals)))